News Articles Summarizer/streamlit application/requirements.txt
streamlit
newspaper3k
nltk

